deployment/requirements.txt
fastapi 
uvicorn 
pydantic 
openai-whisper 
ffmpeg
python-multipart
gtts
ultralytics 
opencv-python 
numpy 
pillow